move client/selfservice to nutanix/sdks/v3/
